requirements.txt
discord.py==0.16.12
Django==2.0
requests==2.18.4
EsiPy==0.3.3
django-bootstrap4==0.0.4
